% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijks.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm for finding shortest paths between source node and all other nodes in a given graph.}
\source{
https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm.
}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A graph represented in the form of data frame with 3 columns namely v1,v2,w which represents starting node values,ending node values,weight of the path.}

\item{init_node}{A source node in the form of scalar from which shortest distance to every node is calculated.}
}
\value{
A vector containing shortest distance from source node to all other nodes.
}
\description{
The algorithm checks for indices of minimum distance and updates the distance value of its neighbours only if there exists a path from source to current node such that path distance is less than current distance value.
}
