% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3file.R
\name{dijsktra}
\alias{dijsktra}
\title{Dijkstra algorithm for finding shortest paths between source node and all other nodes in a given graph
The function takes two arguments 1:graph in the form of data frame which has three variables v1,v2,w representing starting node,ending node,weight for the path
The algorithm checks for indices of minimum distance and updates the distance value of its neighbours only if there exists a path from source to current node such that path distance is less than current distance value}
\usage{
dijsktra(graph, init_node)
}
\value{
A vector containig shortest distance from source node to all other nodes
https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm
}
\description{
Dijkstra algorithm for finding shortest paths between source node and all other nodes in a given graph
The function takes two arguments 1:graph in the form of data frame which has three variables v1,v2,w representing starting node,ending node,weight for the path
The algorithm checks for indices of minimum distance and updates the distance value of its neighbours only if there exists a path from source to current node such that path distance is less than current distance value
}
