% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm to find greatest common divisor.}
\source{
https://en.wikipedia.org/wiki/Euclidean_algorithm.
}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{a numeric scalar.}

\item{y}{another numeric scalar.}
}
\value{
The greatest common divisor of two numbers passed in as arguments.
}
\description{
The algorithm replaces larger of the two numbers by the remainder when divided by smaller of two numbers and stops when remainder equals zero.
}
