% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_lab3.R
\docType{package}
\name{r_lab3}
\alias{r_lab3}
\alias{r_lab3-package}
\title{r_lab3: A package for computating the greatest common divisior of two numbers.}
\description{
The r_lab3 package provides two important functions:
euclidean and dijkstra.
}
\section{r_lab3 functions}{
 
The euclidean function is to find the greatest common divisor of two numbers.
}

