% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki graph data
#' @format A data frame with 18 rows and 3 columns
\describe{
\item{v1}{index of the source vertices}
\item{v2}{index of the destination vertices}
\item{w}{cost of the path taken}
}}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
wiki graph data
#' @format A data frame with 18 rows and 3 columns
\describe{
\item{v1}{index of the source vertices}
\item{v2}{index of the destination vertices}
\item{w}{cost of the path taken}
}
}
\keyword{datasets}
