% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3file.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm to find greatest common divisor
The function takes two numbers for which GCD will be calculated as arguments
The algorithm replaces larger of the two numbers by the remainder when divided by smaller of two numbers and stops when remainder equals zero}
\usage{
euclidean(x, y)
}
\value{
The greatest common divisor of two numbers passed in as arguments
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
\description{
Euclidean algorithm to find greatest common divisor
The function takes two numbers for which GCD will be calculated as arguments
The algorithm replaces larger of the two numbers by the remainder when divided by smaller of two numbers and stops when remainder equals zero
}
